Vendor gems for the demo app
As a follow up to https://github.com/primer/view_components/pull/170,
this change vendors gems when the demo app gets set up.

This change allows for `./script/setup` to run without needing the user
to fill out a password if the home dir is protected. This should make
setup easier for some users.

The PVC dep also got bumped, which seems like a good thing.