Add static dump of component statuses
We'd like a way to inspect the statuses of components in projects where
this library is used without having to load the entire library (which
requires loading Rails and can be slow for things like linters).

This PR adds some code and a rake task to dump the components' statuses
as JSON into a file that can be read and parsed without loading the
environment.

We also add specs to make sure the file is up to date when developers
change the status of a component. If they change a status and don't
update this file, CI will fail and tell them to run the rake task and
commit the diff.