Reduce allocations in Primer::BaseComponent test selector
When we initialize Primer::BaseComponent, we are calling `except` on the
args hash twice. Once is to remove a test selector and the other is to
remove classify related keys.

This change experiments with calling except only once. We allocate a new
array to generate the list to exclude but should allocate only one
additional hash due to except instead of two.