Add API for restricting system arguments with helpful suggestions  (#634)
### Problem

In pairing with @scotttesler we ended up with a `p: 4` on a BorderBox, an option we should not allow to be set on BorderBox. Only by reading the docs did we find the `padding` option that accomplished what we were trying to do.

### Proposed solution

Introduce a way to restrict which system arguments can be passed to a Primer ViewComponent, starting with padding arguments for `BorderBox`. I've set us up with a nice error message:

```console
ArgumentError: Found 1 violations:

The p system argument is not allowed here. Perhaps you could consider using :padding options of :default, :condensed, and :spacious?
```