Autocorrect components with known SystemArguments (#713)
With the `classes_to_hash` utility, we can now autocorrect components that use known system arguments, such as `mr-1`, `p-3`, etc..
This will expand the coverage of our autocorrectors and will also automatically update them when we migrate more utility classes to use the `utilities.yml`.

To simplify the implementation (and DRY), I moved most of the logic to `Base`. Now each mapper will only have to implement a `class -> args` mapper and a mapper of a specific set of HTML attributes (defined by the ATTRIBUTES constant).
Moving everything to `Base` makes it easier to reuse logic like the SystemArguments conversion and `aria- / data-` arguments.