Don't convert arguments with ERB blocks (#712)
Currently some linters can incorrectly convert values that have erb blocks. E.g.:

`<span class="Label" title="<%= some_erb %>">` becomes `Primer::LabelComponent.new(title: "<%= some_erb %>")` which does not interpolate the value (it should be using `#{}`).
There's a similar problem with `ClipboardCopy#value` and the general `data-test-id`.

In this PR, I'm extracting a helper to identify when a value has an erb block and blocking the conversion if that's the case.