Add autocorrection for CloseButton (#770)
The `CloseButton` linter will analyze the content from the HTML tag and decide whether or not it should autocorrect it.
It will only correct an offense if the `close-button` element's content is a single `octicon`, `primer_oction`, or `Primer::OcticonComponent`. It will also check if either the element or its content has an `aria-label` and use it in the correction.

To be able to analyze an element's content, this PR changes how the `tag_tree` is built. Before, it would only iterate over HTML tags, but now it will analyze all elements of the AST (HTML, text, erb...).