Stabilize Ruby 3 allocations
I don't have a great understanding of this, but `GC.compact` likely
invalidates caches which seem to be causing our Ruby 3.0 allocations to
appear larger than they actually are.

This change removes the call to `GC.compact` which drops our allocation
numbers. I think the new numbers should be closer to what we expect when
using these classes in an application.