Make octicon cache key construction more performant (#790)
We currently use an intermediate array and string to construct the key for the octicon cache. I tried using `Hash#hash` instead and got almost a 2.5x speedup and a ~16% memory reduction 😄  I saw a nearly 3x speedup without the call to `correct_key_args?`, but it's nice to have that arg safety.

There are two other improvements in here too: 1) passing system arguments one at a time, and compacting/uniquing in-line. These combined with the hashing trick result in a cumulative 24% memory reduction.