Pulling in text color utilities from primer/css and removing functional_text_colors class (#759)
The purpose of the PR is to move color arguments into `Primer::Classify::Utilities` I'm pulling these files directly from primer/css like the rest of the utilities.

One change to the way Utilities work in this PR is now, when an unsupported value in production is used, instead of returning an empty string, I'm returning a dasherized `key-value` class for the argument. This is important as it lets us still support deprecated classes while moving forward with the library.