Replacing dot calls with their reciever code (#821)
In https://github.com/github/github/pull/192446/checks?check_run_id=3680708010 I turned on the rubocop linters in PVC for dotcom. For the `primer_octicon` cop, there was replacements suggested that was catching calls with a `helpers.` defined.

```
C: [Correctable] Primer/PrimerOcticon: Replace the octicon helper with primer_octicon. See https://primer.style/view-components/components/octicon for details.

  helpers.octicon("shield-lock"),
  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

Upon replacement, I found it was replacing all of the call with `primer_octicon`.

```diff
-helpers.octicon("shield-lock")
+primer_octicon("shield-lock")
```
---

In this PR, I added a test for this, and fixed the corrector to take into account `dot?`. I'm not sure if this is the best approach so open to ideas if you know something better.