Lock Node and Yarn version
I couldn't get Storybook to work out of the box because I was using
the latest version of Node, and there's no indication in the codebase
of what version you're expected to be on.

The long-term support version (currently 16.3.0) works as expected.
This locks each package.json to it and the latest version of Yarn,
and by using engine-strict it won't build if you're not on at least
that version.