Fix link to installation doc in RequiresJSFlash
When hosted on primer.style the docs site nests all the pages
under /view-components using Gatsby's pathPrefix functionality.
This installation link was working as expected locally where
the pages are at the root of the server, but on the production
primer.style it was linking to the root index page. By using
Gatsby's Link component it should prefix the path in production
environments.

See:
- https://www.gatsbyjs.com/docs/how-to/previews-deploys-hosting/path-prefix/