Add Spinner system arguments to outermost element
Since the a11y wrapper was added in #845 there's been a report of regression
due to the classes argument being nested in the HTML output. This also shone
a light on the a11y implications of using CSS to show/hide the loader -
because any css triggered by a class added to the Spinner would only effect
the child svg element, it wouldn't show/hide the loading label for
accessibility, so the label would always be read out even if the spinner was
hidden, and wouldn't alert screen reader users when toggled on.

By moving the system arguments to the outermost element the selector issue
is solved as well as the a11y issue.