Update button spacing (#897)
Related to https://github.com/github/primer/issues/253#issuecomment-949215306

I'm changing the `Button` ERB structure so we can avoid unwanted spaces and newlines, which were interfering with spacing. See https://github.com/github/view_component/pull/1121

As stated in the issue, I'm increasing the spacing from icon and counter to 8px and I'm also updating `Dropdown` to use the `Button` caret instead of creating a custom one

Before/After:

![image](https://user-images.githubusercontent.com/11280312/142928588-9f8d30eb-f5a1-4a3f-aa84-f0a254c605a3.png)
![image](https://user-images.githubusercontent.com/11280312/142928614-c7406068-703a-412a-9acf-8c5476c03a49.png)

The change is really subtle, but it does have a slight difference. Before we used margins of 4px but also had a whitespace at the beginning of the text, which made the actual distance of around 6px