Release and publish scripts v2 (#925)
Working with bash is not the ideal dev experience, so I'm moving our scripts to `thor`! Since `thor` scripts are written with Ruby (which most of our team is experienced with), iterating over them should be easier.

I'm also adding new features to the scripts:

1. `release` will create the PR for you
2. `publish` will create the release for you

Note: It's really hard to test this since I don't want to create unnecessary releases, but I'm highly confident in the code. Worst-case scenario we can fix bugs while doing the next release