Trigger workflows on commit created by docs-build action
Events triggered with the default token won't create a new workflow run,
but using the shared Primer token will trigger them.

We want to trigger all workflows when a commit is made for updating
the docs because otherwise it can leave PRs in a permanent waiting
state with required checks not being run. It also can disguise
a PR with failing checks.

See:
- https://github.com/peter-evans/create-pull-request/issues/48
- https://github.com/peter-evans/create-pull-request/blob/v3.12.0/docs/concepts-guidelines.md#triggering-further-workflow-runs