Pass changelog check if the PR has 'skip changelog' label
We want to be able to bypass adding to the Changelog if our PR changes don't
alter the library output. Primer React uses a 'skip changeset' label for
a similar feature; this means that not adding to the Changelog is an
intentional action by contributors rather than being forgotten.

This updates the action triggers to include adding and removing labels from
the PR.

See:
- https://github.com/primer/react/blob/v34.1.0/.github/workflows/check_for_changeset.yml
- https://stackoverflow.com/questions/62325286/run-github-actions-when-pull-requests-have-a-specific-label#comment122159108_62331521
- https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#pull_request
- https://docs.github.com/en/actions/learn-github-actions/expressions#example-using-an-array