Lock demo app Primer CSS version to version defined in package.json
We're using the Rails asset pipeline to load the Primer CSS file. We're
using a post-install script on the package.json to copy it into a local
folder. We thought about interpolating the version into a link from
unpkg.com but that would require JSON parsing and we'd be relying on an
external registry to have the version we requested.

We tried to duplicate this setup for the Storybook layouts (instead of
importing the CSS in preview.js) but had issues emulating the host option
for stylesheet_link_tag. See https://github.com/rails/rails/issues/35364

We also needed to add a yarn install to the CI action for the stylesheet
to be available for the asset pipeline.