Fix ADR navigation (#971)
* Fix sidebar casing of ADRs

This gets the title of each ADR from the page heading instead of
deriving it from the filename. I've removed the ADR number from
the start of each title because a) it's ugly b) it's not part of
standard ADR templates, see https://github.com/joelparkerhenderson/architecture-decision-record

* Add simple ADR page

Currently the sidebar heading for ADRs links to a broken page.
We can't store this file at docs/content/adr/index.md because
the docs:build script empties that directory and programatically
adds ADRs from the root adr directory.

* Hide initial ADR on docs site

It provides useful context for contributors but doesn't belong on
the docs site. Most of the information is also covered in on the
Architecture decisions index page.