Ignore docs-build action if pushed by Dependabot
Dependabot doesn't have access to secrets for security reasons, which
we need here because it uses git write access to push an extra commit.

We could get around this using multiple workflows to narrow-down the
permissions scope specifically for Dependabot but it's messy and has
its own security risks.

I think it's unlikely that we'll have a dependabot update that would
change the built docs so I think it's okay for us to skip this check
in these circumstances.

See:
- https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-852541544
- https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-797125425
- https://github.blog/changelog/2021-02-19-github-actions-workflows-triggered-by-dependabot-prs-will-run-with-read-only-permissions/