Add skip-changelog label to Dependabot PRs
Currently Dependabot PRs auto-fail CI because of the Changelog check.
As we're only getting Dependabot PRs for security fixes, @primer/css
and @primer/gatsby-theme-doctocat, they aren't lib changes that would
require a Changelog entry.

It's possible we may have Dependabot PRs in the future that _should_
have Changelog entries so this should be reviewed at a later date,
or we should have guidance on how to review Dependabot PRs that
includes checking whether they effect the library output.