Copy deploy-preview action from @primer/css
This removes the steps for Storybook deployments (something we won't be able to
do in PVC because Storybook relies on a live Rails server) and updates the paths
to use app instead of src (we may want to add demo here as well?). I added the
VERCEL_PROJECT_ID as a repo secret.