Add .vercelignore file
I've indentified that the problem with queued Vercel builds is because the
Vercel CLI which is used by the vercel-action uploads all files aside from those
in the default ignore (https://vercel.com/docs/concepts/deployments/build-step#ignored-files-and-folders),
which only includes the root node_modules.

I've also unchecked a box in the UI called 'Include source files outside of the
Root Directory in the Build Step' to avoid including anything ourside of the
docs folder.

See - https://vercel.com/guides/prevent-uploading-sourcepaths-with-vercelignore