Add workflow for production deploys
This is a copy of the preview deploy workflow but without the deployment status
event that is visible in PRs, and it changes all the 'preview' names to 'production'
and passes the prod flag to Vercel so it uses the production domain.

We need to move the production workflow to an action as well because the now-build
script no longer builts the Gatsby site in the Vercel side; we have to do it in the
action before we trigger the build.

This is an identical setup to @primer/css and should not result in any functional
changes.