Comment out Ruby static file generation
Currently all the static files are committed to the repo and are kept up to date with
the docs-build action, so we don't need to build them as part of this action. I've left
them in here because the intention of #982 is to remove all those files from git and
scrap that action, so they will be needed soon.