Use GitHub actions for Vercel deployments (#989)
* Copy deploy-preview action from @primer/css

This removes the steps for Storybook deployments (something we won't be able to
do in PVC because Storybook relies on a live Rails server) and updates the paths
to use app instead of src (we may want to add demo here as well?). I added the
VERCEL_PROJECT_ID as a repo secret.

* Stop auto-deployment from Vercel bot

* Update deprecated Vercel config

Name is deprecated; version is always 2 so can be omitted. now.json still works
but could be deprecated any day, and all newer projects use vercel.json.

* Build static files before deployment

This frees us up from having to keep these in the file system for deployment.
The steps are copied directly from the existing docs-build workflow.

* Build docs site in action

Vercel installs the dependencies and runs the build npm script during its
deployment, but we can possibly speed things up by doing that on our end and
only deploying the static results. This may require settings changes that are
only available in the Vercel UI so can't be versioned.

* Add .vercelignore file

I've indentified that the problem with queued Vercel builds is because the
Vercel CLI which is used by the vercel-action uploads all files aside from those
in the default ignore (https://vercel.com/docs/concepts/deployments/build-step#ignored-files-and-folders),
which only includes the root node_modules.

I've also unchecked a box in the UI called 'Include source files outside of the
Root Directory in the Build Step' to avoid including anything ourside of the
docs folder.

See - https://vercel.com/guides/prevent-uploading-sourcepaths-with-vercelignore

* Stop building Gatsby on Vercel

The Vercel build step is defined in the settings UI so we can't turn it off
because it will effect current production builds as well. It does run the
now-build script though, and we can shortcircuit this to a successful exit
immediately.

* Start pending deployment at start of workflow

* Remove unnecessary parameters

* Add workflow for production deploys

This is a copy of the preview deploy workflow but without the deployment status
event that is visible in PRs, and it changes all the 'preview' names to 'production'
and passes the prod flag to Vercel so it uses the production domain.

We need to move the production workflow to an action as well because the now-build
script no longer builts the Gatsby site in the Vercel side; we have to do it in the
action before we trigger the build.

This is an identical setup to @primer/css and should not result in any functional
changes.

* Comment out Ruby static file generation

Currently all the static files are committed to the repo and are kept up to date with
the docs-build action, so we don't need to build them as part of this action. I've left
them in here because the intention of #982 is to remove all those files from git and
scrap that action, so they will be needed soon.