Remove preview deployments for Dependabot-generated PRs
Since #989 we've been using a GitHub action for deployment of preview
apps and production. The preview action fails on Dependabot PRs because
Dependabot-run actions have read-only GitHub tokens and don't have
access to secrets, so they can't initiate the Deployment action step.

This turns off the deploy-preview workflow for actions initiated by
Dependabot so the whole CI doesn't fail (the same solution as #980).

We _should_ have preview deployments for dependency PRs but it's almost
impossible with this actions workflow. In this repo all PRs have to be
up to date with the base branch before merging though, so reviewers are
often forced to click 'update branch' manually before merging, which
will trigger a deploy because it triggers a merge by the user who
clicked the button and therefore that user triggers the actions. This
workflow can be used by reviewers who need to check the docs site for
changes. It's not ideal but preferable to having checks fail on all
dependency PRs.

For more info on the secrets issue see:
- https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-852541544
- https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-797125425
- https://github.blog/changelog/2021-02-19-github-actions-workflows-triggered-by-dependabot-prs-will-run-with-read-only-permissions/