Remove generated docs content from repo (#996)
* Add docs:build docs content to .gitignore

* Remove ignored docs content from git history

With `git rm -r docs/content/adr/ docs/content/components/`

* Generate static files in deployment actions

* Remove forced git-add of docs in docs-build workflow

* Add static file generation to setup script

With the git-ignore of content changes, running script/dev on a fresh
install was breaking because the doctocat Gatsby build is initiated
before the generation of the markdown files. I'm not sure whether we
could modify the Procfile to wait until the initial docs build is
complete before running doctocat, but that would slow down every run
of script/dev; it's neater to run this in the setup script as a one-
time event.