PR Suggestions (#1044)
* Inline `#addEvents` into `connectedCallback`

Since the method is only called in `connectedCallback` we can inline it
there and remove a check for `this.control`.

* Remove internal state getters
These should be exposed via attributes on the element if we want to
expose them but I think we just want them to be part of the internal
API.

* Move `observedAttributes` next to the `attributeChangedCallback`

This makes it easier to understand what attributes we care about and how
the component responds to those attributes

* Group component properties together

* Move TOOLTIP_OFFSET to the only function it's used in

We don't have to keep it in the global scope because it's not global.

* Use TOOLTIP_OPEN_CLASS variable in CSS

* Group event handling callbacks together

* Move single getter below getter and setter pairs

* Remove DIRECTION_TO_CLASS

Co-authored-by: Kate Higa <16447748+khiga8@users.noreply.github.com>