Prevent tooltip from sometimes changing button type
Due to how component `content` is rendered, if a tooltip slot is
evaluated before the button is rendered its `type` will be set to
`:description`, which is not a valid button type. This is due to trying
to set the tooltip type by modifying an instance variable.

This resolves the issue by not overriding `type` and instead setting it
explicitly on the arguments passed to the tooltip.