Prevent slot from overriding component system_arguments (#1199)
* Prevent tooltip from sometimes changing button type

Due to how component `content` is rendered, if a tooltip slot is
evaluated before the button is rendered its `type` will be set to
`:description`, which is not a valid button type. This is due to trying
to set the tooltip type by modifying an instance variable.

This resolves the issue by not overriding `type` and instead setting it
explicitly on the arguments passed to the tooltip.

* Don't allow tooltip to overwrite link system arguments

In scenarios where the `content` of a link component is evaluated before
it's rendered the tooltip slot will overwrite the arguments of the link
component.

This resolves the issue by not assigning @system_arguments within the
slot.