ClipboardCopy only needs aria-label when empty
We were requiring `aria-label` on ClipboardCopy in all cases, but when
the component already contains suitable text this is not recommended.

This changes when we validate the presence of `aria-label`. At
instantiation time the component doesn't know if it will be rendered
with a block, so we check at render time instead.

Fixes #1081