`ClipboardCopy` only needs `aria-label` when empty (#1244)
* ClipboardCopy only needs aria-label when empty

We were requiring `aria-label` on ClipboardCopy in all cases, but when
the component already contains suitable text this is not recommended.

This changes when we validate the presence of `aria-label`. At
instantiation time the component doesn't know if it will be rendered
with a block, so we check at render time instead.

Fixes #1081

* All ClipboardCopy without content needs aria-label

I was just checking for a block to decide if ClipboardCopy needed an
aria-label. But @camertron suggested that checking for blank content
would be better, and I think the code is nicer as well.

* Fill my intentionally empty block for RuboCop

An explicit nil tests just as well while satisfying the linter.