Bump demo app dev Ruby version from 2.7.1 to 2.7.6
Mostly this is because I ran into problems compiling 2.7.1 on an Apple
Silicon Mac.

This doesn't affect what's running in production (which is already set
up to use the latest 2.7.x), but it is also not a bad idea to have the
latest security patches even when you're not serving to the internet.