Allow form field names to end in "?"
Rails will allow input names that end in a question mark and call the
appropriate method on the model to (for example) decide if a checkbox
should start off as "checked". This change ensures we can still find a
caption template associated with a field named with a question mark, but
not have to embed a question mark in the template file name.

Co-Authored-By: Cameron Dutro <camertron@gmail.com>