Fix a test that fails in non-UTC TZs
This test was passing in CI (which must be in UTC), but when run on
developer machines in other time zones it would fail.

To fix it, we changed the test from making a Time object and converting
to UTC to have it make a Time object in UTC from the beginning.