Reserve space for ToggleSwitch octicon
This will occupy the same amount of space regardless of if we are
showing one of the status octicons (loading or error) to prevent
excessive layout jumping on the page.

Also adds a gap style to the flexbox container to keep a little
distance from the label/caption/validation message.