Fix toggle switch form templates
The ToggleSwitchForm is unique in that we expect developers to use it
directly sometimes but also inherit from it. When we implement the
`inherited` method on ToggleSwitchForm we also need to call super to
allow that same method on Primer::Forms::Base to do its magic as well
(in this case, finding related templates).

Co-Authored-By: Cameron Dutro <camertron@gmail.com>