Fix tooltips opening when focus is removed while displaying
This is an issue when, for example, a dialog with a tooltip on the show
button closes and the focus changes, because the tool-tip element opens
and doesn't get hidden when the focus moves away.

This fixes that by registering a global focusin listener to check after
a brief period of time if the showing tooltip was focussed into is still
focussed. If it is not then it gets hidden.

This in theory would create a small flash of the tooltip but our
animations on tooltips gives us enough buffer so that the tooltip isn't
visible right away.