Allow setting an `id` attribute on the BorderBox's `ul` element
In order to perform turbo stream append and prepend actions on a
`BorderBox` component, we need a way to specify a target via an `id`
(or some unique CSS selector). We can't guarantee there is a unique
selector for the specific BorderBox's `ul` element we want to append
to without an id.

This commit allows providing the `id` that will be rendered in the
`BorderBox`'s ul element via the `list_id` system argument.

A system argument has been added to the "Playground" preview for the
`BorderBox`.