fix layering issues with dialog/popover hack
While the prior fix in `6cc198512c` worked to keep the dialog open,
because it called showPopover to open the ancestral popovers, they ended
up on the top of the stack, rather than underneath the dialog. This
means while the dialog was open, in certain resolutions it can be
occluded by its ancestral popups.

This new fix instead changes the popovers to be manually controlled,
meaning they won't participate in the "hide all popovers" algo, meaning
they won't close when the dialog is open. When the dialog closes we turn
them back onto auto so that they can continue to have the auto
behaviours such as click-outside, escape, etc.