Emit an input event primer-text-field#clearContents.
Having the `#clearContents` emit an `input` event, it integrates better
whenever an `input` event listener is bound to the `primer-text-field`
components. A classic example is to use the `primer-text-field` as a
search field, which triggers a background request when the field's value
is changed. Without emiting the `input` event on #clearContents, the
filtering background request would not be triggered.