Do not call actions within assert_selector block
(tldr: not to be confused with `within`)

Only other assertions should be called within `&optional_filter_block`:
this filter block is passed to `Capybara::SelectorQuery`, which resets
the wait time to 0 before calling it. Actions such as `fill_in` may rely
on a longer wait time - this is the case in particular for our overriden
`fill_in` (see `System::TestCase`), which executes the Axe accessibility
testing engine asynchronously after a text field is filled in.