Fix DangerDialog body scroll behaviour
Applies `display: contents` style to `danger-dialog-form-helper`, which
makes the custom element behave as if it doesn't exist in the DOM
structure, passing all styles directly to its children.

This makes `scrollable-region` behaves correcly, which in turn ensures
the dialog behaves the same as `Primer::Alpha::Dialog` with regards to
scrolling.